requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pygraphviz==1.14
PyJWT==2.9.0
sqlparse==0.5.3
